OpenBSD local modifications: - privilege separation - some snprintf/strlc{at,py} conversions - support for more architectures - signal handler fixes - default mouse wheel support in config tools - sunmouse protocol

+++ xserver/hw/xfree86/common/xf86Xinput.c
@@ -950,7 +950,7 @@ xf86PostMotionEvent(DeviceIntPtr	device,
 			}
 		    }
 		    else if (dx || dy) {
-			mult = pow((float)(dx*dx+dy*dy),
+			mult = pow((float)dx*(float)dx + (float)dy*(float)dy,
 				   ((float)(device->ptrfeed->ctrl.num) /
 				    (float)(device->ptrfeed->ctrl.den) - 1.0) / 
 				   2.0) / 2.0;
